Separate mapper tests
